feat: :sparkles: add user management functionalities; implement user retrieval, update, and guild member fetching; integrate winston for logging; adjust ability values and refactor imports
