+++ frontend/package.json
@@ -21,7 +21,8 @@
     "next-auth": "^5.0.0-beta.25",
     "react": "^18.3.1",
     "react-d3-tree": "^3.6.2",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "winston": "^3.17.0"
   },
   "devDependencies": {
     "@types/node": "^22.9.0",
